Deslocando a classe Animate para Arquivo animação
